refactor(servicenow)!: rename resolver FQCN to com.keepersecurity.secretsManager.CredentialResolver
Move the resolver and its KeeperCredentialMapper helper (and test) from
com.snc.discovery.keeper into Keeper's namespace com.keepersecurity.secretsManager,
renaming the class to CredentialResolver. The pre-FQCN shared class
com.snc.discovery.CredentialResolver keeps its MID-mandated name and now subclasses the
renamed class by FQN. Updates the jar Main-Class.

Also gate FQCN registration to Yokohama (Patch 7+): the External Credential Resolver FQCN
field is available from Yokohama Patch 7+, so Xanadu and older ship the legacy shared class,
while Yokohama, Zurich and Australia ship the FQCN-only jar. Updates the publish-workflow
variant case, build.gradle.kts comments, README (prose + compatibility matrix) and the
unreleased 1.0.0 CHANGELOG.

Refs: KSM-1093
