+++ integration/servicenow-external-credential-resolver/build.gradle.kts
@@ -69,17 +69,18 @@ tasks.named<org.cyclonedx.gradle.CycloneDxTask>("cyclonedxBom") {
 }
 
 // Resolver JAR variant (pass with -PresolverVariant=fqcn|legacy):
-//   fqcn   -> ServiceNow Xanadu+ : ships the unique com.snc.discovery.keeper.KeeperCredentialResolver
-//             and OMITS the shared com.snc.discovery.CredentialResolver, so Keeper coexists with
-//             other vendors' resolvers (which ship that shared class) on the same MID Server.
-//   legacy -> pre-Xanadu : keeps com.snc.discovery.CredentialResolver (required there).
+//   fqcn   -> ServiceNow Yokohama (Patch 7+) and newer : ships the unique
+//             com.keepersecurity.secretsManager.CredentialResolver and OMITS the shared
+//             com.snc.discovery.CredentialResolver, so Keeper coexists with other vendors'
+//             resolvers (which ship that shared class) on the same MID Server.
+//   legacy -> Xanadu and older : keeps com.snc.discovery.CredentialResolver (required there).
 // Default "legacy" keeps both classes (also convenient for local dev/testing).
 val resolverVariant = (project.findProperty("resolverVariant") as String? ?: "legacy").lowercase()
 
 tasks.jar {
     duplicatesStrategy = DuplicatesStrategy.EXCLUDE
     manifest {
-        attributes("Main-Class" to "com.snc.discovery.keeper.KeeperCredentialResolver")
+        attributes("Main-Class" to "com.keepersecurity.secretsManager.CredentialResolver")
     }
     from(configurations
         .runtimeClasspath
@@ -90,8 +91,8 @@ tasks.jar {
     exclude("META-INF/*.SF")
     exclude("META-INF/*.DSA")
     exclude("META-INF/*.RSA")
-    // Xanadu+ variant drops ONLY the shared legacy class (com.snc.discovery.CredentialResolver) so
-    // it doesn't clash with other vendors' JARs; the unique com.snc.discovery.keeper.* class stays.
+    // The fqcn variant drops ONLY the shared legacy class (com.snc.discovery.CredentialResolver) so
+    // it doesn't clash with other vendors' JARs; the unique com.keepersecurity.secretsManager.* class stays.
     if (resolverVariant == "fqcn") {
         exclude("com/snc/discovery/CredentialResolver.class")
     }
