+++ integration/servicenow-external-credential-resolver/src/test/java/com/keepersecurity/secretsManager/KeeperCredentialMapperTest.java
@@ -1,4 +1,4 @@
-package com.snc.discovery.keeper;
+package com.keepersecurity.secretsManager;
 
 import com.keepersecurity.secretsManager.core.HiddenField;
 import com.keepersecurity.secretsManager.core.KeeperRecord;
